[core] extend dictionary module to support UID dictionaries
- move data element (tag) type definitions to `dictionary::data_element`
- add `dictionary::uid` and respective types
